requirements.txt
openai==1.59.9
streamlit==1.40.0
ipykernel==6.29.4
langchain==0.1.16
langchain-community==0.0.34
langchain-openai
pypdf==4.2.0
faiss-cpu==1.8.0
python-dotenv
unidecode
pydub
ffmpeg
moviepy
requests-toolbelt